\name{acmacs-package}
\alias{acmacs-package}
\alias{acmacs}
\docType{package}
\title{
  low level access to the data stored in acmacs files
}
\description{
  Reading/writing acmacs and lispmds files, accessing data stored in those files
}
\details{
}
\author{
Eugene Skepner, eu@antigenic-cartography.org.

Maintainer: Eugene Skepner <eu@antigenic-cartography.org>
}
\keyword{ package }
\seealso{
}
\examples{
  \dontrun{
     ## Optional simple examples of the most important functions
     ## These can be in \dontrun{} and \donttest{} blocks.
  }
}
