\name{acmacs.Antigen}
\alias{acmacs.Antigen-class}
\alias{acmacs.Antigen}
\docType{package}
\title{Class to acces data antigen related data.}
\usage{
antigen <- chart$antigens[[antigen_number]]
antigen$name
antigen$passage
antigen$reassortant
}
\value{An object of class \code{"acmacs.Antigen"} to access and modify
  antigen data. It has the following properties:

  \item{$name}{name of the antigen.}
  \item{$passage}{}
}
\keyword{ package }
\author{Eugene Skepner, eu@antigenic-cartography.org.}
